requirement_gpu.txt
absl-py==0.3.0
astor==0.7.1
beautifulsoup4==4.6.0
cycler==0.10.0
gast==0.2.0
google==2.0.1
grpcio==1.13.0
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.15.0
pandas==0.23.3
pkg-resources==0.0.0
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
Werkzeug==0.14.1
scikit-learn
PyDrive
